// src/lib/weebdex/model/mangaListItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Suicaodex API
 * WeebDex API đè tem, ăn quả úp bô nên không backup kịp data.

Phải gì ạ? Phải chịuuuuuu....
 * OpenAPI spec version: 1.0.0
 */
import type { MangaListItemAltTitles } from "./mangaListItemAltTitles";
import type { MangaRelationships } from "./mangaRelationships";

export interface MangaListItem {
  /** @nullable */
  alt_titles?: MangaListItemAltTitles;
  /** @nullable */
  content_rating?: string | null;
  created_at?: string;
  /** @nullable */
  demographic?: string | null;
  /** @nullable */
  description?: string | null;
  id: string;
  language?: string;
  /** @nullable */
  last_chapter?: string | null;
  /** @nullable */
  last_uploaded_chapter_at?: string | null;
  /** @nullable */
  last_volume?: string | null;
  published_at?: string;
  relationships?: MangaRelationships;
  state?: string;
  /** @nullable */
  status?: string | null;
  title: string;
  updated_at?: string;
  /** @nullable */
  version?: number | null;
  /** @nullable */
  year?: number | null;
  [key: string]: unknown | null;
}